pyproject.toml
[project]
name = "clap"
version = "0.1.0"
description = "CLAP: Clinical LLM Audit Pack - Release-governance harness for Clinical LLMs"
readme = "README.md"
requires-python = ">=3.11"
license = { text = "Apache-2.0" }
authors = [{ name = "CLAP Contributors" }]
keywords = ["clinical", "llm", "audit", "evaluation", "synthetic"]
classifiers = [
    "Development Status :: 3 - Alpha",
    "Intended Audience :: Science/Research",
    "License :: OSI Approved :: Apache Software License",
]

dependencies = [
    "pyyaml>=6.0",
    "jsonschema>=4.20",
    "reportlab>=4.0",
    "matplotlib>=3.7",
    "pandas>=2.0",
    "numpy>=1.24",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.4",
    "pytest-cov>=4.1",
]

[project.scripts]
clap = "clap.cli:main"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["clap"]

[tool.hatch.build.targets.sdist]
include = ["clap", "data", "experiments", "paper"]

[tool.pytest.ini_options]
testpaths = ["tests"]
pythonpath = ["."]
